main service NFT action GRPC msg
